refactor: split bot.py monolith (1557 lines) into modular handlers
Closes #17, resolves #22 (composition.py decoupled from app.bot)

- Created app/handlers/ with 10 focused modules (all ≤ 206 lines)
- formatting.py, process_runners.py, actions.py, registry.py
- agents.py, agents_discovery.py, terminal.py, chat.py
- auth.py, project.py, approval.py, commands.py, delegation.py
- bot.py trimmed from 1557 → 206 lines (core handlers only)
- composition.py no longer imports from app.bot — uses handlers.registry
  and handlers.approval directly
